sendmail

+++ TaskMail/Controllers/SendMailController.cs
@@ -25,7 +25,7 @@ namespace TaskMail.Controllers
         [HttpGet]
         public IActionResult TaskMail([Required]long headerId, [Required]long userId)
         {
-            _SendMail.TaskMail(headerId, userId, out int _status, out string _message, out int _mailCount);
+            _SendMail.TaskMail(headerId, userId, out int _status, out string _message, out long _mailCount);
             return StatusCode(CommonDetails.StatusCode(_status), new { data = new { }, status = _status, message = _message , mailCount=_mailCount});
         }
